pyproject.toml
[project]
name = "lingbot-map"
version = "0.1.0"
description = "LingBot-Map: Geometric Context Transformer for Streaming 3D Reconstruction"
requires-python = ">= 3.10"
dependencies = [
    "Pillow",
    "huggingface_hub",
    "einops",
    "safetensors",
    "opencv-python",
    "tqdm",
    "scipy",
    "torchvision",
]

[project.optional-dependencies]
vis = ["viser>=0.2.23", "trimesh", "matplotlib", "onnxruntime", "requests"]
live = ["fastapi>=0.115.0", "uvicorn[standard]>=0.32.0", "websockets>=13.0"]
demo = ["lingbot-map[vis]"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lingbot_map*"]
